project/requirements.txt
asgiref==3.8.1
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.43.0
Django==5.2
django-cloudinary-storage==0.3.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
iniconfig==2.1.0
Markdown==3.7
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.5
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
